requirements.txt
# Core dependencies
opencv-python>=4.8.0
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

# Google integrations (optional)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gspread>=5.11.0

# Optional: Advanced facial recognition
# face_recognition>=1.3.0  # Uncomment if you want high-accuracy facial recognition

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
